Rename TorchScript compiler to IR emitter to better reflect its function.
[ghstack-poisoned]
