+++ tools/build_variables.py
@@ -151,7 +151,7 @@ libtorch_sources = [
     "torch/csrc/jit/register_special_ops.cpp",
     "torch/csrc/jit/register_distributed_ops.cpp",
     "torch/csrc/jit/scope.cpp",
-    "torch/csrc/jit/script/compiler.cpp",
+    "torch/csrc/jit/script/iremitter.cpp",
     "torch/csrc/jit/script/edit_distance.cpp",
     "torch/csrc/jit/script/logging.cpp",
     "torch/csrc/jit/script/convert_to_ssa.cpp",
