+++ torch/csrc/jit/autodiff.cpp
@@ -1,6 +1,7 @@
 #include <torch/csrc/jit/autodiff.h>
 
 #include <ATen/core/functional.h>
+#include <c10/util/Exception.h>
 #include <torch/csrc/jit/jit_log.h>
 #include <torch/csrc/jit/operator.h>
 #include <torch/csrc/jit/passes/common_subexpression_elimination.h>
@@ -8,9 +9,8 @@
 #include <torch/csrc/jit/passes/dead_code_elimination.h>
 #include <torch/csrc/jit/passes/inliner.h>
 #include <torch/csrc/jit/passes/lower_tuples.h>
-#include <torch/csrc/jit/script/compiler.h>
+#include <torch/csrc/jit/script/iremitter.h>
 #include <torch/csrc/jit/symbolic_script.h>
-#include <c10/util/Exception.h>
 
 #include <algorithm>
 #include <memory>
