+++ torch/jit/supported_ops.py
@@ -161,7 +161,7 @@ def _get_math_builtins():
 
 
 def _get_global_builtins():
-    # Taken from the 'globals' map in torch/csrc/jit/script/compiler.cpp
+    # Taken from the 'globals' map in torch/csrc/jit/script/iremitter.cpp
     supported_builtins = [
         'print',
         'tuple',
